Fix spacing below search in mastery view
When you have a large number of standard so the table is off the screen using
`margin` doesn't seem to work. Switching to use `padding` works in both the case
where the table is larger then the viewable screen and when it is within the
contents on the screen.

Tested with Chrome, Firefox and Edge. Tested that stick header and first column
are unaffected.