US127502 - Allow Removal of Ratio
This allows the removal of the ratio in the top left as well as not displaying
the outdated icon. Layout and CSS were adjust to allow for the complete removal
of the ratio without disrupting the layout as well as fixing some vertical
alignment issues. Also fixed up some issues with demo data.

Tested in **Chrome**, **Firefox** and **Edge** using dev BSI and demo pages.