US127501 - Allow removal of the histogram
This allows the removal of the histogram at top of each column as well as
disables the calls the back-end for data retrieval. Layout and CSS were adjust
to allow for the complete removal of the histogram without disrupting the
layout. A new attribute (`disable-graph`) was introduced on the table view
component that can be added to the component at render time to globally disable
the histograms.

Tested in **Chrome**, **Firefox** and **Edge** using dev BSI and demo pages.